feat(test): 픽스처 생성
